Fix for race condition with external code.
Alteration to how the queue clearing system works to both make it
more flexible and also to eliminate a race condition with external
code. Rather than being a single function that causes the remaining
queue contents to be returned, clearing now causes the queue's
lower limit on content to hit zero causing a consumer to empty the
queue. This can then by synchronised correctly in external code and
not leave the queue waiting for more messages when there will not be
any.

Signed-off-by: Thomas Bytheway <tb403@cam.ac.uk>